% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A Hello Function}
\usage{
hello(really = T)
}
\arguments{
\item{really}{Should you really print it ? Defaults to TRUE.}
}
\description{
This function allows you to say Hello
}
\examples{
hello()
}
\keyword{hello}

