% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kaggle_titanic_test}
\alias{kaggle_titanic_test}
\title{Survival on the Titanic}
\format{A data table with 418 rows and 11 variables:}
\source{
\url{https://www.kaggle.com/c/titanic/data}
}
\usage{
kaggle_titanic_test
}
\description{
A dataset containing the Titanic's passengers's data, and their survival.
}
\keyword{datasets}
