% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlprepr.R
\docType{package}
\name{mlprepr}
\alias{mlprepr}
\alias{mlprepr-package}
\title{mlprepr: Prepare data for machine learning, and save this preparation}
\description{
The mlprepr package provides three categories of important functions:
learn, drift and apply.
}
\section{Learn functions}{

The learn functions ...
}

