% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_winsor.R
\name{winsor_learn}
\alias{winsor_learn}
\title{A function to learn the winsorization of a column}
\usage{
winsor_learn(col.val, percent.min = 5/100, percent.max = 95/100)
}
\arguments{
\item{col.val}{the column to winsorize}

\item{percent.min}{the quantile used for winsorization, defaults to 5%}

\item{percent.max}{the quantile used for winsorization, defaults to 95%}
}
\value{
a list with two values called "min" and "max"
}
\description{
This function allows you to learn the winsorization of a column
}
\examples{
WinsorLearn(col, 0.05, 0.95)
}
\keyword{data.table}
\keyword{winsor}
