% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_winsor.R
\name{winsor_predict}
\alias{winsor_predict}
\title{A function to apply an existing winsorization to a column}
\usage{
winsor_predict(col.val, col.nam, min, max)
}
\arguments{
\item{col.val}{the column to winsorize}

\item{min}{the minimum value allowed. Those which are below are replaced.}

\item{max}{same as min, but above max value.}

\item{the}{name of the column to create}
}
\value{
a data.table with the new column
}
\description{
This function allows you to apply an existing winsorization to a column
}
\examples{
WinsorLearn(col, "mycolname", list(min = 0.05, max = 0.95))
}
\keyword{data.table}
\keyword{winsor}
